cylc/cylc#887: update isodatetime
